Use more explitit types

+++ python/basic/Operators.C
@@ -75,8 +75,8 @@ struct OperatorInt {
 
 struct OperatorLong {
 #ifdef _WIN64
-   operator intptr_t() { return m_long; }
-   intptr_t m_long;
+   operator int64_t() { return m_long; }
+   int64_t m_long;
 #else
    operator long() { return m_long; }
    long m_long;
@@ -100,8 +100,8 @@ struct OperatorUnsignedInt {
 
 struct OperatorUnsignedLong {
 #ifdef _WIN64
-   operator uintptr_t() { return m_ulong; }
-   uintptr_t m_ulong;
+   operator uint64_t() { return m_ulong; }
+   uint64_t m_ulong;
 #else
    operator unsigned long() { return m_ulong; }
    unsigned long m_ulong;
